% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{as_database}
\alias{as_database}
\title{Decompose one or more obpg filenames into the obpg database}
\usage{
as_database(
  x = c("A2004001.L3m_DAY_CHL_chlor_a_4km.nc", "A2004001.L3m_DAY_SST_sst_4km.nc",
    "AQUA_MODIS.20191130.L3m.DAY.NSST.sst.4km.NRT.nc",
    "NPP_VIIRS.20190703.L3m.DAY.SST.sst.4km.nc")
)
}
\arguments{
\item{x}{character, vector of filenames to decompose. Path is discarded.}
}
\value{
a database as a data frame (tibble)
 \itemize{
   \item{date, Date  We only use L3 which is by day, so no POSIXct}
   \item{year, character date formatted as YYYY}
   \item{mmdd, character date formatted as mmdd}
   \item{mit, character mission-instrument-type}
   \item{level, character string indicating the level}
   \item{period, charcater period indicator for L3}
   \item{suite, character suite identifier}
   \item{param, character parameter name chlor_a, sst, etc}
   \item{per, compositing period such as DAY, MO, 8D}
   \item{res, character resolution in km}
   \item{nrt, character optional Near Real-Time identifier}
   \item{file, basename of the input file less the extension}
 }
}
\description{
Decompose one or more obpg filenames into the obpg database
}
