% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{as_filename}
\alias{as_filename}
\title{Convert a database entry to a filename}
\usage{
as_filename(x, path = ".", ext = ".tif")
}
\arguments{
\item{x}{data frame (tibble), to convert}

\item{path}{character, the path to prepend to the filename}

\item{ext}{character, the extension to append to the filename}
}
\value{
one or more filenames
}
\description{
Convert a database entry to a filename
}
