% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoy.R
\name{buoy_locations}
\alias{buoy_locations}
\title{Retrieve buoy location information}
\usage{
buoy_locations()
}
\value{
a data frame (tibble) of ID, lon and lat
}
\description{
Retrieve buoy location information
}
