% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{invert_vector}
\alias{invert_vector}
\title{Exchange a character vector values and names}
\usage{
invert_vector(x)
}
\arguments{
\item{x}{a named character vector}
}
\value{
a named character vector
}
\description{
Exchange a character vector values and names
}
