% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_obpg1}
\alias{is_obpg1}
\title{Determine if a filename is in OPBPG old-style format ("opbg1")}
\usage{
is_obpg1(
  x = c("A2004001.L3m_DAY_CHL_chlor_a_4km.nc", "A2004001.L3m_DAY_SST_sst_4km.nc",
    "AQUA_MODIS.20191130.L3m.DAY.NSST.sst.4km.NRT.nc",
    "NPP_VIIRS.20190703.L3m.DAY.SST.sst.4km.nc", "Ocean Hack Week")
)
}
\arguments{
\item{x}{character a vector of one or more filenames (with or without path)}
}
\value{
logical TRUE where the naming pattern matches the old-style OBPG naming convention
}
\description{
Based upon the assumption the old-style uses up to 3 characters to identify
the platform before staring with the date.
}
