% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{obpg1_to_obpg2}
\alias{obpg1_to_obpg2}
\title{Convert OBPG old-style filenames to OBPG new-style filenames}
\usage{
obpg1_to_obpg2(
  x = c("A2004001.L3m_DAY_CHL_chlor_a_4km.nc", "A2004001.L3m_DAY_SST_sst_4km.nc",
    "AQUA_MODIS.20191130.L3m.DAY.NSST.sst.4km.NRT.nc",
    "NPP_VIIRS.20190703.L3m.DAY.SST.sst.4km.nc", "Ocean Hack Week")
)
}
\arguments{
\item{x}{character, one or more obpg style filenames,
directory paths and extensions are dropped}
}
\value{
character vector of obpg2 style filenames without path and extension
}
\description{
A2004001.L3m_DAY_CHL_chlor_a_4km becomes AQUA_MODIS.20040101.L3m.DAY.CHL.chlor_a.4km
}
