% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{obpg_base_url}
\alias{obpg_base_url}
\title{Retrieve the base url for either HTML or OPENDAP}
\usage{
obpg_base_url(type = c("html", "opendap")[2])
}
\arguments{
\item{type}{character, either "html" or "opendap" (default)}
}
\value{
URL
}
\description{
Retrieve the base url for either HTML or OPENDAP
}
