% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{obpg_build_url}
\alias{obpg_build_url}
\title{Build a path for date, level, suite, resolution and parameter}
\usage{
obpg_build_url(
  dates = as.Date("2018-01-01"),
  lvl = "L3m",
  param = "chlor_a",
  suite = "CHL",
  period = "DAY",
  res = "4km",
  platform = "AQUA_MODIS",
  baseuri = obpg_base_url(type = "opendap"),
  ext = ".nc"
)
}
\arguments{
\item{dates}{Date, one or more dates as character (YYYY-mm-dd), Date-class, POSIXt-class
If requesting a periods other than DAY, then dates refers to the start of the period}

\item{lvl}{character, the processing level Currently only "L3m" is tested.}

\item{param}{character, the name of parameter}

\item{suite}{character, the name of the paramter suite}

\item{period}{character, the period the imagery covers ala (DAY, 8D, MO, ...). Currently only
DAY has been tested.}

\item{res}{charcater, either '4km' (default) or '9km'}

\item{platform}{character, the platform (mission, instrument and platform).  Currently, only
"AQUA_MODIS" has been tested}

\item{baseuri}{character, the base URL to the resource}

\item{ext}{character, an optional extension to append to the URL}
}
\value{
character URL
}
\description{
Build a path for date, level, suite, resolution and parameter
}
