% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{obpg_close}
\alias{obpg_close}
\title{Close an ncdf4 connection to an OBPG URL}
\usage{
obpg_close(x)
}
\arguments{
\item{x}{ncdf4 class object}
}
\value{
NULL invisibly
}
\description{
Close an ncdf4 connection to an OBPG URL
}
