% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{obpg_nc_nav}
\alias{obpg_nc_nav}
\title{Retrieve OBPG navigation values (start, count, lons, lats).  Create one of these
to resuse hwen downloading a suite of images from related NCDF resources (like a series
of days for a given product and parameter).}
\usage{
obpg_nc_nav(
  x,
  bb = c(-77, -42.5, 36.5, 56.7),
  res = obpg_res(what = "4km"),
  varname = "sst"
)
}
\arguments{
\item{x}{ncdfd4 object}

\item{bb}{numeric, 4 element requested bounding box [west, east, south, north]}

\item{res}{numeric, 2 element resolution [res_x,res_y]}

\item{varname}{character the name of the variable}
}
\value{
list with
\itemize{
  \item{bb the requested bounding box}
  \item{res the resolution}
  \item{start vector of start values}
  \item{count vector of count values}
  \item{ext vector of extent (for raster::raster)}
  \item{crs character, proj string for raster::raster}
  \item{varname character}
}
}
\description{
Retrieve OBPG navigation values (start, count, lons, lats).  Create one of these
to resuse hwen downloading a suite of images from related NCDF resources (like a series
of days for a given product and parameter).
}
