% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{obpg_open}
\alias{obpg_open}
\title{Open an ncdf4 connection to an OBPG URL}
\usage{
obpg_open(uri, try_nrt = TRUE, verbose = FALSE)
}
\arguments{
\item{uri}{character the URL}

\item{try_nrt}{logical, if TRUE and the connection fails, then try again
after modifying the url to ".NRT.nc" instead of ".nc"}

\item{verbose}{logical, if TRUE voice warnings}
}
\value{
either ncdf4 object or NULL
}
\description{
Open an ncdf4 connection to an OBPG URL
}
