% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{obpg_period_string}
\alias{obpg_period_string}
\title{Given a start date and a period code and style generate a an OBPG date pairing}
\usage{
obpg_period_string(dates, period = "DAY", style = "new")
}
\arguments{
\item{dates}{Date-class the dates (always refers to the start if an extended period)}

\item{period}{character the period of the product (DAY, 8D, MO, etc)}

\item{style}{character either "new" or "old"}
}
\value{
character vector with either YYYYmmdd or YYYYjjj style format
}
\description{
Given a start date and a period code and style generate a an OBPG date pairing
}
