% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{obpg_res}
\alias{obpg_res}
\title{Retrieve the known OBPG resolution}
\usage{
obpg_res(what = c("4km", "9km")[1])
}
\arguments{
\item{what}{character, specify the resultion with either "4km" or "9km"}
}
\value{
a two element [x,y] vector
}
\description{
Retrieve the known OBPG resolution
}
