% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{obpg_style}
\alias{obpg_style}
\title{Guess the OBPG filenaming style from the parameter name}
\usage{
obpg_style(x = c("chlor_a", "par", "sst", "sst4"))
}
\arguments{
\item{x}{one or more parameters}
}
\value{
character vector, one per input, of "new" or "old" indicating filenaming style
}
\description{
Guess the OBPG filenaming style from the parameter name
}
