% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{platform_codes}
\alias{platform_codes}
\title{Returns a named list of known platform codes}
\usage{
platform_codes(invert = FALSE, version = c("2", "1")[1])
}
\arguments{
\item{invert}{logical if TRUE invert the name-value relationship}

\item{version}{character either "2" (default) or "1".  Refers to the
filenaming OBPG convention}
}
\value{
a named character vector of platform codes
}
\description{
Returns a named list of known platform codes
}
