% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{read_database}
\alias{read_database}
\title{Read a database file stored as CSV (possible compressed)}
\usage{
read_database(path, filename = "database.csv.gz")
}
\arguments{
\item{path}{character, the path to the database}

\item{filename}{character, by default 'database.csv.gz' but it can be any other filename}
}
\description{
Read a database file stored as CSV (possible compressed)
}
