% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{read_navigation}
\alias{read_navigation}
\title{Read the navigation data structure relative to world}
\usage{
read_navigation(x = c("world_4km", "world_9km")[1])
}
\arguments{
\item{x}{character, the name of the navigation file to read (default = "world_4km")}
}
\value{
a list of navigation elements
\itemize{
  \item{bb the requested bounding box}
  \item{res the resolution as from \code{\link[raster]{res}}}
  \item{start vector of start values \code{[column, row]}}
  \item{count vector of count values \code{[columns, rows]}}
  \item{ext vector of extent as from \code{\link[raster]{extent}}}
  \item{crs character, proj string as from \code{\link[raster]{projection}}}
  \item{varname character}
}
}
\description{
Read the navigation data structure relative to world
}
