% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoy.R
\name{read_buoy}
\alias{read_buoy}
\title{Read in buoy data}
\usage{
read_buoy(buoy = "M01")
}
\arguments{
\item{buoy}{the}
}
\description{
Read in buoy data
}
\seealso{
\href{http://gyre.umeoce.maine.edu/data/gomoos/buoy/html/M01.html}{UMaine Buoy Data}
}
